% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_result_class.R
\name{plot.ml_result_class}
\alias{plot.ml_result_class}
\title{Plot an ml_result_class object}
\usage{
\method{plot}{ml_result_class}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ml_result_class}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a confusion matrix heatmap.
}
